services/presentation/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-pptx==0.6.23
pydantic==2.6.4
